reorganizing filter_set file location
