+++ lib/cuke_slicer/filters/filter_set.rb
